# specs/002-the-api-app/contracts/omni-secrets.openapi.yaml
openapi: 3.1.0
info:
  title: Apex Omni Secret Management API
  version: 0.1.0
  description: |
    Internal endpoints that orchestrate Google Secret Manager usage for Apex Omni credentials.
    All routes require operator or service authentication with scoped IAM roles.
servers:
  - url: https://api.apex-tradebill.com
    description: Production
  - url: https://staging-api.apex-tradebill.com
    description: Staging (non-sensitive test credentials only)
components:
  securitySchemes:
    serviceAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    SecretType:
      type: string
      enum: [trading_api_key, trading_client_secret, webhook_shared_secret]
...
paths:
...
  /internal/apex-omni/secrets/cache/refresh:
    post:
      summary: Force the API process to evict cached secrets and fetch from GSM
      security:
        - serviceAuth: []
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                secretType:
                  $ref: '#/components/schemas/SecretType'
                  description: Refresh a specific secret; refreshes all when omitted
      responses:
        '202':
          description: Refresh scheduled
          content:
            application/json:
              schema:
                type: object
                properties:
                  requestId:
                    type: string
                  refreshedSecretTypes:
                    type: array
                    items:
                      $ref: '#/components/schemas/SecretType'
...